backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.32.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-storage-blob==12.24.0
azurefunctions-extensions-base==1.0.0b2
azurefunctions-extensions-bindings-blob==1.0.0b3
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.1
packaging==24.2
pathspec==0.12.1
pip-autoremove==0.10.0
platformdirs==4.3.6
portalocker==2.10.1
propcache==0.3.1
pycparser==2.22
pygrib==2.1.6
PyJWT==2.10.1
pyproj==3.7.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
surfpy @ git+https://github.com/mpiannucci/surfpy.git@c2c4b288777ec539dfa07aa9aeda51a68edcfdd4
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
